Update README.md
Favorite Characters: Ponyo, Sosuke, Mother of Ponyo
Description: Ponyo is a movie from ghibli studio. Ponyo is a fish who fall inlove to the idea of being a human and living as a human.
Who would I recommend it to: I would recommend it to other people who likes to watch ghibli movies.